requirements.txt
pandas>=2.2
PyYAML>=6.0
matplotlib>=3.8


